+++ R/DMR4glm.R
@@ -63,7 +63,7 @@ DMR4glm <- function(X, y, clust.method, lam){
         heig <- c()
         models <- list()
     }
-    len <- length(heig)
+
     heig <- c(0,heig)
     names(heig)[1] = "full"
     if ((p.fac + 1) < p){