+++ R/DMR4lm_help.R
@@ -48,7 +48,6 @@ DMR4lm_help <- function(X, y, clust.method, lam){
         heig <- c()
         models <- list()
     }
-    len <- length(heig)
     heig <- c(0,heig)
     names(heig)[1] = "full"
     if ((p.fac + 1) < p){
@@ -95,16 +94,17 @@ DMR4lm_help <- function(X, y, clust.method, lam){
            if(length(sp[[kt]][sp[[kt]] != 1]) > 0){
                                        sp[[kt]][sp[[kt]] != 1] <- sp[[kt]][sp[[kt]] != 1] + min(spold[spold != 1]) - min(sp[[kt]][sp[[kt]] != 1])
            }
-           ii <- min(which(spold != sp[[kt]]))
-           suma <- ifelse(kt == 1, 0, sum(n.levels[1:(kt-1)] - 1))
-           if(sp[[kt]][ii] == 1){
-             a[suma + ii] <- 1
-           } else {
-             a[suma + ii] <- 1
-             a[suma + min(which(sp[[kt]] == sp[[kt]][ii]))] <- -1
+           for (ii in min(which(spold != sp[[kt]]))) {
+             suma <- ifelse(kt == 1, 0, sum(n.levels[1:(kt-1)] - 1))
+             if(sp[[kt]][ii] == 1){
+               a[suma + ii] <- 1
+             } else {
+               a[suma + ii] <- 1
+               a[suma + min(which(sp[[kt]] == sp[[kt]][ii]))] <- -1
+             }
+             if (kt < length(sp)) for( x in (kt+1):length(sp)){ if (length(sp[[x]][sp[[x]]!=1]) > 0 ) sp[[x]][sp[[x]]!= 1] = sp[[x]][sp[[x]]!=1] - 1}
+             nl <- nl - 1
            }
-           if (kt < length(sp)) for( x in (kt+1):length(sp)){ if (length(sp[[x]][sp[[x]]!=1]) > 0 ) sp[[x]][sp[[x]]!= 1] = sp[[x]][sp[[x]]!=1] - 1}
-           nl <- nl - 1
          }
          A <- cbind(A, a)
          be <- c(0, 2:(p-i+2))