+++ R/glamer_4lm_help.R
@@ -4,6 +4,6 @@ glamer_4lm_help <- function(S, betas_with_intercept, X, y, fl, clust.method, lam
     return(list(b = c(1, rep(0, sum(fl-1))), rss = sum(mm$res^2)))
   }
 
-  mfin <- clusters_4lm_help(S, betas_with_intercept, X, y, clust.method, lam)
+  mfin <- clusters1D_4lm_help(S, betas_with_intercept, X, y, clust.method, lam)
   return(mfin)
 }